requirements.txt
absl-py==2.0.0
asgiref==3.7.2
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
constantly==23.10.4
cryptography==41.0.5
daphne==4.0.0
Django==4.2.7
django-admin-soft-dashboard==1.0.12
flatbuffers==23.5.26
gast==0.5.4
google-auth==2.23.4
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.2
h11==0.14.0
h5py==3.10.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
joblib==1.3.2
keras==2.14.0
libclang==16.0.6
lxml==4.9.3
Markdown==3.5.1
MarkupSafe==2.1.3
ml-dtypes==0.2.0
msgpack==1.0.7
mysqlclient==2.2.0
numpy==1.26.1
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.2
Pillow==10.1.0
plotly==5.18.0
protobuf==4.25.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.3.0
PyPDF2==3.0.1
python-dateutil==2.8.2
python-docx==1.1.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
uvicorn==0.24.0.post1
Werkzeug==3.0.1
wrapt==1.14.1
zope.interface==6.1